Move in_jupyter_notebook function
